_ARCHIVE_TO_SORT/legacy_backends/backend-python/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=21.2.0

# CLI
click>=8.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9

# Environment and Config
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0

# Monitoring
psutil>=5.9.0

# Redis for caching
redis>=5.0.0

# Background tasks
celery>=5.3.0

# HTTP requests
requests>=2.32.0
httpx>=0.27.0

# Date/Time
python-dateutil>=2.9.0

# CORS
fastapi-cors>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# WebSockets
websockets>=12.0
